requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4